requirements.txt
Flask==0.6
Jinja2==2.5.2
Werkzeug==0.6.2
eventlet==0.9.12
feedparser==4.1
feedservice==0.1
greenlet==0.3.1
gunicorn==0.11.0
pymongo==1.8.1
python-dateutil==1.5
wsgiref==0.1.2
